requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
pandas>=2.0.0
numpy>=1.24.0
xgboost>=2.0.0
scikit-learn>=1.3.0
pydantic>=2.0.0
joblib>=1.3.0
shap>=0.44.0
google-generativeai>=0.8.0
python-dotenv>=1.0.1
